requirements.txt
rasa-nlu
jsonlib
pyTelegramBotAPI
Pillow
